task completed

+++ src/main/java/com/wagawin/demo/repository/ParentSummaryRepository.java
@@ -1,9 +1,12 @@
 package com.wagawin.demo.repository;
 
+import java.util.stream.Stream;
+
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.wagawin.demo.model.ParentSummary;
 
-public interface ParentSummaryRepository extends PagingAndSortingRepository<ParentSummary, Long> {
+public interface ParentSummaryRepository extends PagingAndSortingRepository<ParentSummary, Integer> {	
+	
 	  
 }
